Refactor shared folder into component folder. Add basic validation rule to form.
