+++ src/components/FeedbackItem.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { FaTimes } from 'react-icons/fa';
-import Card from '../shared/Card'
+import Card from './shared/Card'
 
 const FeedbackItem = ({ item, onDeleteFeedbackItem }) => {
