+++ src/components/shared/Button.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 const Button = ({ children, type = 'button', version = 'primary', isDisabled = false}) => {
   return (
-    <button type={type} isDisabled className={`btn btn-${version}`}>
+    <button type={type} disabled={isDisabled} className={`btn btn-${version}`}>
       { children }
     </button>
   )
